research_report/requirements.txt
agno 
streamlit
python-dotenv
requests
